backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database - PostgreSQL
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Database - MongoDB
motor==3.3.2
pymongo==4.6.0

# Cache/Queue - Redis
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# AI/LLM
groq==0.4.2

# Utilities
python-dotenv==1.0.0
click==8.1.7

# Logging
python-json-logger==2.0.7

# CORS
fastapi-cors==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.0
flake8==6.1.0

# Optional: File Upload
python-multipart==0.0.6

# Optional: SMS (Twilio)
twilio==8.10.0

# Optional: Email
aiosmtplib==3.0.0
jinja2==3.1.2
